chaining split arrows spawn at correct heights and point the right way
