+++ packages/client/src/app/3d-world/game-world/replay-tree-manager/entity-motion-update-handlers/handle-entity-motion-update.ts
@@ -44,6 +44,11 @@ export function handleEntityMotionUpdate(
       motionUpdate.entityId
     );
 
+    if (motionUpdate.despawn) {
+      actionEntityModelOption.cleanup({ softCleanup: true });
+      return;
+    }
+
     if (motionUpdate.setParent !== undefined)
       handleEntityMotionSetNewParentUpdate(actionEntityModelOption, motionUpdate.setParent);
