+++ packages/client/src/app/3d-world/game-world/replay-tree-manager/entity-motion-update-handlers/handle-start-pointing-toward.ts
@@ -21,5 +21,6 @@ export function handleStartPointingTowardEntity(
     targetPosition
   );
 
-  sceneEntity.movementManager.startRotatingTowards(newRotation, duration, () => {});
+  if (duration === 0) sceneEntity.movementManager.transformNode.rotationQuaternion = newRotation;
+  else sceneEntity.movementManager.startRotatingTowards(newRotation, duration, () => {});
 }
