+++ packages/common/src/action-entities/index.ts
@@ -2,7 +2,10 @@ export * from "./cosmetic-effect.js";
 export * from "./cosmetic-effect-constructors.js";
 import { Vector3 } from "@babylonjs/core";
 import { EntityProperties } from "../primatives/index.js";
-import { SceneEntityChildTransformNodeIdentifier } from "../scene-entities/index.js";
+import {
+  SceneEntityChildTransformNodeIdentifier,
+  SceneEntityChildTransformNodeIdentifierWithDuration,
+} from "../scene-entities/index.js";
 
 export enum ActionEntityName {
   Arrow,
@@ -21,9 +24,11 @@ export const ACTION_ENTITY_STRINGS: Record<ActionEntityName, string> = {
 export type ActionEntityProperties = {
   position: Vector3;
   name: ActionEntityName;
+  initialCosmeticYPosition?: SceneEntityChildTransformNodeIdentifier;
   parentOption?: SceneEntityChildTransformNodeIdentifier;
   initialRotation?: Vector3;
   initialPointToward?: SceneEntityChildTransformNodeIdentifier;
+  initialLockRotationToFace?: SceneEntityChildTransformNodeIdentifierWithDuration;
 };
 
 export class ActionEntity {
