+++ packages/common/src/app-consts.ts
@@ -68,7 +68,7 @@ export const COMBATANT_POSITION_SPACING_BETWEEN_ROWS = 5.0;
 export const COMBATANT_TIME_TO_MOVE_ONE_METER = 300;
 // export const COMBATANT_TIME_TO_MOVE_ONE_METER = 100;
 // export const ARROW_TIME_TO_MOVE_ONE_METER = 40;
-export const ARROW_TIME_TO_MOVE_ONE_METER = 200;
+export const ARROW_TIME_TO_MOVE_ONE_METER = 65;
 export const COMBATANT_TIME_TO_ROTATE_360 = 1000;
 export const MISSING_ANIMATION_DEFAULT_ACTION_FALLBACK_TIME = 1000;
 export const DEFAULT_HITBOX_RADIUS_FALLBACK = 1.5;
